backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-q2==1.7.3
django-cors-headers==4.4.0
